Add chunk file
