+++ docs/usage/images.md
@@ -8,8 +8,8 @@ Most of the recently released [Mistral models](https://huggingface.co/mistralai/
 ## Supported image formats
 
 Mistral Image encoders use Pillow to decode images and OpenCV to encode. Hence, the supported formats are the same as Pillow's. The images can be provided as:
-- an [ImageURLChunk][mistral_common.protocol.instruct.messages.ImageURLChunk]: a pydantic model containing an image URL from which the image will be downloaded.
-- an [ImageChunk][mistral_common.protocol.instruct.messages.ImageChunk]: a pydantic model containing a serialized image that can be either a base64 string or a pillow image.
+- an [ImageURLChunk][mistral_common.protocol.instruct.chunk.ImageURLChunk]: a pydantic model containing an image URL from which the image will be downloaded.
+- an [ImageChunk][mistral_common.protocol.instruct.chunk.ImageChunk]: a pydantic model containing a serialized image that can be either a base64 string or a pillow image.
 
 ## Use an Image encoder with our tokenizer
 
@@ -22,7 +22,7 @@ The attributes of the [ImageConfig][mistral_common.tokens.tokenizers.image.Image
 - `spatial_merge_size`: the number of patches to merge into one token. This is useful to reduce the number of redundant tokens in the image. E.g if the image is 224x224 and the patch size is 14, then the image will be divided into 16x16 patches. If the spatial merge size is 2, then the image will be divided into 8x8 patches.
 
 ```python
-from mistral_common.protocol.instruct.messages import ImageURLChunk
+from mistral_common.protocol.instruct.chunk import ImageURLChunk
 from mistral_common.tokens.tokenizers.image import ImageEncoder, ImageConfig, SpecialImageIDs
 
 special_ids = SpecialImageIDs(img=10, img_break=11, img_end=12)  # These are normally automatically set by the tokenizer
@@ -42,7 +42,8 @@ Let's load the tekken tokenizer used for [Mistral Small 3.1's](https://mistral.a
 ```python
 from huggingface_hub import hf_hub_download
 
-from mistral_common.protocol.instruct.messages import ImageURLChunk, TextChunk, UserMessage
+from mistral_common.protocol.instruct.chunk import ImageURLChunk, TextChunk
+from mistral_common.protocol.instruct.messages import UserMessage
 from mistral_common.protocol.instruct.request import ChatCompletionRequest
 from mistral_common.tokens.tokenizers.mistral import MistralTokenizer
